refactor: 변경된 엔티티에 맞춰 발주 생성 dto를 변경한다.
